refactor: organise interfaces
